requirements.txt
Django==4.2.2
django-debug-toolbar==4.1.0
django-import-export==3.2.0
pandas==2.0.2
matplotlib==3.7.1